- Remove feedback and settings for teacher profile - Merge about fragment and profile fragment - Add sign out button to profile
