+++ app/src/main/res/menu/profile_menu.xml
@@ -1,6 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <menu xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto">
+
     <item
         android:id="@+id/profile_setting_item"
         android:icon="@drawable/ic_settings_black_24dp"
