+++ app/src/main/res/navigation/t_main_navigation.xml
@@ -2,7 +2,7 @@
 <navigation xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
-    android:id="@+id/main_navigation"
+    android:id="@+id/t_main_navigation"
     app:startDestination="@id/homeFragment">
 
     <fragment
@@ -23,7 +23,7 @@
     <action
         android:id="@+id/action_global_homeFragment"
         app:destination="@id/homeFragment"
-        app:popUpTo="@id/main_navigation"
+        app:popUpTo="@id/t_main_navigation"
         app:popUpToInclusive="true" />
     <fragment
         android:id="@+id/myCourseFragment"
@@ -41,13 +41,13 @@
     <action
         android:id="@+id/action_global_myCourseFragment"
         app:destination="@id/myCourseFragment"
-        app:popUpTo="@id/main_navigation"
+        app:popUpTo="@id/t_main_navigation"
         app:popUpToInclusive="true" />
     <fragment
         android:id="@+id/profileFragment"
         android:name="com.mobile.ta.teacher.view.user.profile.ProfileFragment"
         android:label="@string/toolbar_profile"
-        tools:layout="@layout/frag_profile">
+        tools:layout="@layout/t_frag_profile">
         <action
             android:id="@+id/action_profileFragment_to_editProfileFragment"
             app:destination="@id/editProfileFragment" />
@@ -58,7 +58,7 @@
     <action
         android:id="@+id/action_global_profileFragment"
         app:destination="@id/profileFragment"
-        app:popUpTo="@id/main_navigation"
+        app:popUpTo="@id/t_main_navigation"
         app:popUpToInclusive="true" />
     <fragment
         android:id="@+id/searchFragment"
@@ -193,20 +193,6 @@
             android:name="user"
             app:argType="com.mobile.ta.model.user.User" />
     </fragment>
-    <fragment
-        android:id="@+id/settingsFragment"
-        android:name="com.mobile.ta.teacher.view.user.settings.SettingsFragment"
-        android:label="@string/toolbar_settings"
-        tools:layout="@layout/fragment_settings">
-        <action
-            android:id="@+id/action_settingsFragment_to_feedbackFragment"
-            app:destination="@id/feedbackFragment" />
-    </fragment>
-    <fragment
-        android:id="@+id/feedbackFragment"
-        android:name="com.mobile.ta.teacher.view.user.feedback.FeedbackFragment"
-        android:label="@string/toolbar_feedback"
-        tools:layout="@layout/fragment_feedback" />
     <fragment
         android:id="@+id/courseSubmitFragment"
         android:name="com.mobile.ta.teacher.view.course.chapter.assignment.CourseSubmitFragment"
